+++ config/routes.rb
@@ -1,3 +1,0 @@
-EmberCli::Engine.routes.draw do
-  get ":app_name", to: "ember_tests#index", format: false
-end