+++ lib/ember-cli/ember_controller.rb
@@ -1,0 +1,9 @@
+module EmberCli
+  class EmberController < ::ApplicationController
+    def index
+      @app = params[:ember_app]
+
+      render layout: false
+    end
+  end
+end